maint: avoid new warning-errors from gcc-6.1
When configured with --enable-gcc-warnings and gcc-6.1, ...
* src/safe.c (count_path_components): Use _GL_ATTRIBUTE_PURE,
to avoid an error from -Werror=suggest-attribute=pure.
* src/util.h (filename_is_safe): Likewise.
* src/patch.c (main): Placate -Werror=format= by casting
pch_mode's mode_t return type to the "unsigned int" required
to match the %o format string.
* src/patch.c (delete_files): Correct indentation, to avoid
this error from -Werror=misleading-indentation:

  patch.c: In function 'delete_files':
  patch.c:1816:4: error: this 'if' clause does not guard...
      if (verbosity == VERBOSE)
      ^~
  patch.c:1820:6: note: ...this statement, but the latter is
    misleadingly indented as if it is guarded by the 'if'
	move_file (0, 0, 0, file_to_delete->name, mode,
	^~~~~~~~~

+++ src/patch.c
@@ -241,8 +241,8 @@ main (int argc, char **argv)
 	{
 	  say ("File %s: can't change file type from 0%o to 0%o.\n",
 	       quotearg (inname),
-	       pch_mode (reverse) & S_IFMT,
-	       pch_mode (! reverse) & S_IFMT);
+	       (unsigned int) (pch_mode (reverse) & S_IFMT),
+	       (unsigned int) (pch_mode (! reverse) & S_IFMT));
 	  skip_rest_of_patch = true;
 	  somefailed = true;
 	}
@@ -1817,9 +1817,9 @@ delete_files (void)
 	    say ("Removing %s %s\n",
 		 S_ISLNK (mode) ? "symbolic link" : "file",
 		 quotearg (file_to_delete->name));
-	    move_file (0, 0, 0, file_to_delete->name, mode,
-		       file_to_delete->backup);
-	    removedirs (file_to_delete->name);
+	  move_file (0, 0, 0, file_to_delete->name, mode,
+		     file_to_delete->backup);
+	  removedirs (file_to_delete->name);
 	}
     }
   gl_list_iterator_free (&iter);
